deepscan-backend/ml_server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
tensorflow==2.19.0
torch>=2.2.0
torchvision>=0.17.0
pytorchcv==0.0.74
h5py>=3.11.0
numpy==2.1.3
Pillow==11.2.1
opencv-python-headless==4.11.0.86
